requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.2
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
discord.py==1.7.1
ffmpeg-python==0.2.0
future==0.18.2
idna==3.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
pycparser==2.20
pylint==2.7.4
PyNaCl==1.4.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.4.7
